Initial ftest1 Based on fdev1
This is to be the production (or at least test without build tools)
environment corresponding to fdev1.  Take the current fdev1 repository
files as a starting point and commit them.

The follow on steps will be to have fdev1 based on ftest12 and install
the gcc tools ontop.  Production libraries go into ftest1, development
libraries and tools go into fdev1.