Change Packages To Those Needed For Tests
The majority of packages installed for fdev1 are not needed in the
production environment.  Reduce the list to just those needed for
running binaries build with the libraries involved in fdev1.

Extending tafthorne/netcat-debian as it is a helpful generic base image
that include netcat, which we require for some of the basic network
smoke test.