Add Protocol Buffer & gRPC
Add gRPC and ProtoBuf libraries to the image.  When used in concert
these should be generated from the gRPC repository as discussed on
https://github.com/grpc/grpc/blob/v1.15.0/BUILDING.md

This is based off of the current fdev1 work.  Hopefully the Docker magic
will mean most of the builder image is reused from cache based on some
finger print.  I will look into whether it would be better to either:
Have the builder image in a seperate Docker file that fdev1 and ftest1
reference
or have a single large Docker file for the fdev1 and ftest1 images.